+++ NEWS.md
@@ -1,3 +1,8 @@
+# ptstem 0.0.2
+
+* Add the complete argument. Before it was used undirectly.
+* Add tests for raising errors
+
 # ptstem 0.0.1
 
 * CRAN release